requirements.txt
openai~=1.6.1
Flask~=3.0.0
python-dotenv
flask-cors
gunicorn
